+++ tools/agic/versions.tf
@@ -3,7 +3,7 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">= 2.51"
+      version = "~> 3.0"
     }
     # tflint-ignore: terraform_unused_required_providers
     helm = {